Highlight emergent behavior in swarms
Hey Nathan! One thought about the description from http://swarmtheplanet.wtf/:

>The Swarm is an homage to the natural truth that the whole is made up of many parts. Mass is a collection of atoms; a network is a collection of clients; a community is a collection of individuals. Embrace these definitions as you consider your Hackathon project.

In my study of networks in school, an important theme is how the whole can be more than the sum of its parts in many systems including social networks. For the sake of accuracy & to further inspire people, I would add a sentence to that paragraph about emergent behavior & how the whole can take on a life much bigger than that of the many parts.

Here’s a quote from the link that Brian Gruber sent about the original model of swarming behavior:
>"In the boids model (and related systems like the multi-agent steering behavior demos) interaction between simple behaviors of individuals produce complex yet organized group behavior. The component behaviors are inherently nonlinear, so mixing them gives the emergent group dynamics a chaotic aspect. At the same time, the negative feedback provided by the behavioral controllers tends to keep the group dynamics ordered. The result is life-like group behavior.” (http://www.red3d.com/cwr/boids/)